Add Resources
Fonts have been added to www/css/fonts for a better cleaner font
than that provided by ionic.

The default avatar has been added to www/img/. This avatar
(www/img/avatar_1.png) is to be shown when there is no profile
picture. This image will be replaced with a better sized image.